requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.3
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
yarl==1.8.1
